% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMeasurement.R
\name{extractMeasurement}
\alias{extractMeasurement}
\title{call measurement data i neede to calculate}
\usage{
extractMeasurement(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  measurementConceptIdSet
)
}
\arguments{
\item{measurementConceptIdSet}{vector of measurement concept ids I need}
}
\description{
call measurement data i neede to calculate
}
